+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - add transformation for imitation
+### Changed
+- use structural pattern matching at applying transformations
 
 ## [0.2.0] - 2024-02-11
 ### Added